front/requirements.txt
Flask
google-cloud-texttospeech